requirements.txt
asgiref==3.8.1
attrs==23.2.0
coverage==7.4.4
dj-database-url==2.1.0
Django==5.0.4
django-mathfilters==1.0.0
django-simple-history==3.5.0
djangorestframework==3.15.1
djangorestframework-api-key==3.0.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.5.1
factory-boy==3.3.0
Faker==24.11.0
gunicorn==22.0.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
packaging==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
whitenoise==6.6.0
